RX DMA Calypte [FEATURE]: add performance counters for blocking behaviour
The performance counters are measuring the amount of empty clock periods
when some data are ready and are waiting for a response/accept by the
receiver. Three counters have been implemented:

1) The relative amount of idle clock periods where the PCIe IP is not able to
receive data that are ready on the RX DMA controller's output, e.g. the
DST_RDY on the MFB bus between the RX DMA and the PCIe IP is deasserted.
2) The relative amount of idling clock periods before receiving a valid address
on which the DMA Header can be written.
3) The relative amount of idling clock periods before receiving a valid address
on which the DATA can be written.
4) The sum of requests for the PCIE address for the DATA transaction.
5) The sum of requests for the PCIE address for the DMA HEADER transaction.
6) The sum of clock periods where data are being transmitted
on the MFB bus towards the PCIe IP

Additionally, two histograms have been created where each bin is a
counter for a channel within the controller. These count the amount of
buffer fills where either DATA buffer or DMA HDR buffer gets filled and
the controller is not able to write data anymore. This is different
counting than the waiting for the available PCIE address.