feat(TX_MAC_LITE): add support for dropping when the link is down
The goal is to unify the TX MAC behavior on different FPGA chips.
Now, if the ETH link is down, the packets are already dropped in
the TX_MAC_LITE. This is especially useful when running ndp-generate
immediately. In the event of a sudden outage of a heavily loaded link
(full FIFO in TX_MAC_LITE) with Intel FPGAs, theoretically,
the entire pipeline can still stop.