fix(pcie_cq_adapter): fix assignment of FBE and LBE
When two regions are used, there are also two FBEs and two LBEs for them
but their representation as specified in the PCIe IP documentation
differs from the one in the NDK platform. The first spec uses the first FBE
and the first LBE to describe the first TLP in the beat and (if present)
the second FBE and the second FBE to describe the second TLP in the
beat. However, the MFB specification applies its metadata (which FBE and
LBE bits are part of) to each region which does not adhere to the spec
of the PCIE IP since there can be only one packet begin but in the
second region. Using this logic would map the second LBE and FBE to the
metadata of this
packet which is not correct. Only first LBE and FBE needs to be parsed
to the second region's metadata.