fix(tx_dma_calypte): add enable to LBE register when generating BE vector
Otherwise, it would not be taken into account that a gap inside a packet
can occur. This gap occured when the Gen4x16 link on Intel Agilex 7 FPGA
downtrained to Gen3x16. In this case, the P-Tile IP did still run on 400
MHz which obviously creates gaps inside a packet.